+++ src/app/coreFilesUpdater/CoreFilesUpdaterFactory.ts
@@ -5,6 +5,7 @@
 import { BaseCoreFilesUpdater } from "./BaseCoreFilesUpdater";
 import * as semver from "semver";
 import { CoreFilesUpdater_2_11 } from "./coreFilesUpdaters/CoreFilesUpdater_2_11";
+import { CoreFilesUpdater_2_14 } from "./coreFilesUpdaters/CoreFilesUpdater_2_14";
 import { CoreFilesUpdaterNop } from "./coreFilesUpdaters/CoreFilesUpdaterNop";
 
 export class CoreFilesUpdaterFactory {
@@ -16,6 +17,8 @@ export class CoreFilesUpdaterFactory {
             return undefined;
         } else if (semver.eq(currentVersion, "2.12.0")) {
             return new CoreFilesUpdaterNop();
+        } else if(semver.lt(currentVersion, "2.14.0")) {
+            return new CoreFilesUpdater_2_14();
         }
         return undefined;
     }