+++ src/app/manifestGeneration/ManifestVersions.ts
@@ -6,5 +6,6 @@ export enum ManifestVersions {
     v13 = "v1.3",
     v14 = "v1.4",
     v15 = "v1.5",
+    v16 = "v1.6",
     devPreview = "devPreview"
 }